+++ packages/core/src/report/ReportBuilder.ts
@@ -321,6 +321,7 @@ export class ReportBuilder {
             )
         );
 
+        /* v8 ignore if -- @preserve */
         if (!isDebugMode()) {
             Object.assign(temporarilyContainer.style, reportStyles.outOfScreen);
         }